edgar_deploy/requirements.txt
pyhaikunator
pyyaml
google-auth
google-cloud-container
google-cloud-storage
kubernetes
docker
setuptools
